requirements.txt
fudge==1.0.3
httplib2==0.9.1
httpretty==0.8.10
nose==1.3.7
oauth2==1.5.211
oauthlib==0.7.2
pycrypto==2.6.1
PyInstaller==2.1
PyTumblr==0.0.6
requests==2.7.0
requests-oauthlib==0.5.0
scrypt==0.7.1
simplejson==3.8.0
soundcloud==0.4.1
twython==3.2.0
wheel==0.24.0
